Show dates, and allow ordering by them
Display of the "public_timestamp" field on the documents, and add
"button" links to the document list to control the ordering of the
documents.

This is apparently useful for content designers to be able to do.

It looks pretty terrible, but so does the rest of this; it's just
intended to check what's possible, and see if there's a genuine need
being met here.