Fix a bug with single-valued fields
Some fields are returned from rummager as a single value (eg, format)
instead of a list (eg, organisations).  This was breaking the compare
view for single-valued fields, because the value returned was being
iterated.  As a result, before this fix, the compare view for format
listed the characters in the format name, instead of the formats
themselves.