Fix collections in CSV output
In the CSV output, we were trying to return doc.collections, which is
the wrong name and never exists, so the column was always empty.

Instead, we need to return doc.document_collections, and extract just
the "link" value for each item in the list.  This provides the list of
links of collections that each document is in.