Make it easier to filter on multiple field values
Before this change, when you add a filter on, say, organisation, the
list of organisations is updated to only include those which are used in
the resulting documents.  This generally means that you get far fewer
options available.

Instead, change the list of values shown in each filter field to be
based on the current scope except for the filters applied to that field.
This allows multiple values to be chosen.  If multiple values are
chosen, documents are selected which match any of those values.